3414 ability to preserve output order (#3925)
* 3414: Changes on regtest scripts to test in an easier way.

* 3414 migration

* 3414: Backend changes

* 3414: Frontend changes.

* small changes

* 3414: Added small test

* 3414: Working in imb.

* 3414: lint fixes

* 3414: Removed debug line

* 3414: output order unit tests..

* 3414: Added docs.

* preserve order documentation language updates

* preserveOutputOrder

* Revert "preserveOutputOrder"

This reverts commit 389cd2d6cd993b976e5af2793fa65b876805f454.

It turns out BIP 69 actually sorts both inputs and outputs, and this disables
both sorts.

* why are outputs reversed?!?!

* the issue is the database returns outputs in random order

We cannot rely on the order returned by the database. However, we can rely that
they are incremeneted in IDs one after the other. The solution is to sort the
outputs that are returned. I cannot easily use a database sorting method here,
but I can easily use a javascript sorting method. This sorts the outputs in the
correct order. The reason why simulated payments don't have this is that
simulated payments do not insert into the database, and also the preserveOrder
method is not being used in simulated payments.

* remove console.logs

* preserve order in simulated payments too

* remove browser-side console.log

* fix docs

* more docs fixes

Co-authored-by: Ryan X. Charles <ryanxcharles@gmail.com>
Co-authored-by: Ryan X. Charles <ryan@moneybutton.com>